+++ integrated_channels/moodle/exporters/content_metadata.py
@@ -115,9 +115,9 @@ class MoodleContentMetadataExporter(ContentMetadataExporter):
             return int(parse(end_date).replace(tzinfo=timezone.utc).timestamp())
         return None
 
-    def _apply_delete_transformation(self, content_metadata_item):
+    def _apply_delete_transformation(self, metadata):
         """
         Specific transformations required for "deleting/hiding" a course on a Moodle.
         """
-        content_metadata_item['visible'] = 0  #hide a course on moodle - instead of true delete
-        return content_metadata_item
+        metadata['visible'] = 0  # hide a course on moodle - instead of true delete
+        return metadata